+++ CHANGELOG.md
@@ -4,7 +4,7 @@
   * Function `EasyTableRowCursor` to overrides a row cursor.
   * Theme attribute `RowThemeData.cursorOnTapGesturesOnly` to allow cursor without tap gestures.
 * `CellThemeData`
-  * The `ownInputDecoration` attribute has been added to overrides the `InputDecorationTheme` with a custom theme.  
+  * The `overrideInputDecoration` attribute has been added to overrides the `InputDecorationTheme` with a custom theme.  
 
 ## 2.3.0